tools/csv_geo_agent/requirements.txt
adk==0.0.5
aiohappyeyeballs==2.4.6
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
attrs==25.1.0
Authlib==1.5.2
bcrypt==4.3.0
beam-client==0.2.153
beautifulsoup4==4.13.4
beta9==0.1.191
betterproto-beta9==2.0.1
black==25.1.0
boto3==1.37.25
botocore==1.37.25
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-plugins==1.1.1
cligj==0.7.2
cloudpickle==3.1.1
contourpy==1.3.2
croniter==3.0.4
cryptography==44.0.2
cycler==0.12.1
dataclasses-json==0.6.7
Deprecated==1.2.18
deprecation==2.1.0
distro==1.9.0
docstring_parser==0.16
dotenv==0.9.9
fastapi==0.115.12
filelock==3.17.0
filetype==1.2.0
fiona==1.10.1
fireworks-ai==0.15.12
fonttools==4.57.0
frozenlist==1.5.0
fsspec==2025.2.0
geographiclib==2.0
geopandas==1.0.1
geopy==2.4.1
google==3.0.0
google-adk==0.4.0
google-ai-generativelanguage==0.6.17
google-api-core==2.24.2
google-api-python-client==2.169.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.91.0
google-cloud-bigquery==3.31.0
google-cloud-core==2.4.3
google-cloud-resource-manager==1.14.2
google-cloud-secret-manager==2.23.3
google-cloud-speech==2.32.0
google-cloud-storage==2.19.0
google-cloud-trace==1.16.1
google-crc32c==1.7.1
google-genai==1.13.0
google-resumable-media==2.7.2
googleapis-common-protos==1.69.2
gotrue==2.12.0
graphviz==0.20.3
grpc-google-iam-v1==0.14.2
grpcio==1.69.0
grpcio-status==1.71.0
grpclib==0.4.7
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
httpx-ws==0.7.1
huggingface-hub==0.29.1
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.6.1
iniconfig==2.1.0
isort==5.13.2
Jinja2==3.1.5
jiter==0.8.2
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.8
langchain==0.3.19
langchain-community==0.3.18
langchain-core==0.3.47
langchain-experimental==0.3.4
langchain-google-genai==2.1.1
langchain-openai==0.3.7
langchain-text-splitters==0.3.6
langsmith==0.3.11
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.1
mcp==1.7.1
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
networkx==3.4.2
numpy==2.2.3
openai==1.64.0
opentelemetry-api==1.32.1
opentelemetry-exporter-gcp-trace==1.9.0
opentelemetry-resourcedetector-gcp==1.9.0a0
opentelemetry-sdk==1.32.1
opentelemetry-semantic-conventions==0.53b1
orjson==3.10.15
packaging==24.2
pandas==2.2.3
paramiko==3.5.1
pathspec==0.12.1
pgeocode==0.5.0
pillow==11.1.0
platformdirs==4.3.6
pluggy==1.5.0
postgrest==1.0.1
prompt_toolkit==3.0.50
propcache==0.3.0
proto-plus==1.26.1
protobuf==4.25.7
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.8.0
pydantic_core==2.27.2
Pygments==2.19.1
PyJWT==2.10.1
PyNaCl==1.5.0
pyogrio==0.10.0
pyparsing==3.2.3
pyproj==3.7.1
pytest==8.3.5
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2025.1
PyYAML==6.0.2
realtime==2.4.3
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rsa==4.9
s3transfer==0.11.4
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
sentence-transformers==3.4.1
setuptools==75.8.1
shapely==2.1.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.38
sse-starlette==2.3.4
starlette==0.46.2
storage3==0.11.3
StrEnum==0.4.15
supabase==2.15.1
supafunc==0.9.4
sympy==1.13.1
tabulate==0.9.0
tenacity==9.0.0
threadpoolctl==3.5.0
tiktoken==0.9.0
tokenizers==0.21.0
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
tqdm==4.67.1
transformers==4.49.0
typeguard==2.13.3
typing-inspect==0.9.0
typing_extensions==4.13.2
tzdata==2025.1
tzlocal==5.3.1
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.2
watchdog==4.0.2
wcwidth==0.2.13
websocket-client==1.8.0
websockets==13.1
wrapt==1.17.2
wsproto==1.2.0
yarl==1.18.3
zipp==3.21.0
zstandard==0.23.0
